#101 Consolidate sync_universe_config into map_loader, remove redundant file
