+++ src/game.rs
@@ -43,6 +43,7 @@ pub use config::load_map;
 pub use config::load_map_into_db;
 pub use config::load_resources_into_db;
 pub use config::should_auto_load;
+pub use config::sync_universe_config;
 pub use engagement::Engagement;
 pub use engagement::EngagementType;
 pub use engagement::Engagements;
@@ -58,7 +59,6 @@ pub use map::Navigation;
 pub use map::Room;
 pub use map::Universe;
 pub use map::World;
-pub use map::sync_universe_config;
 pub use messaging::PlayerMessage;
 pub use messaging::message;
 pub use player::Player;
