+++ src/game/config/map_loader.rs
@@ -1,9 +1,11 @@
 use sqlx::SqlitePool;
 use std::collections::HashMap;
 use std::error::Error;
+use std::path::Path;
 
 use crate::game::component::Attribute;
-use crate::game::config::entity_config::EntityTypeConfig;
+use crate::game::config::entity_config::{EntityTypeConfig, load_entity_configs};
+use crate::game::config::map_config::load_map;
 use crate::game::config::{FactionConfig, ResourceConfig};
 use crate::game::{EntityConfig, EntityType, Location, Room, Universe, World};
 use crate::persistence::{
@@ -12,6 +14,23 @@ use crate::persistence::{
 };
 
 const LAST_MAP_LOAD_KEY: &str = "last_map_load_date";
+
+pub async fn sync_universe_config(
+    pool: &SqlitePool,
+    config_path: Option<&Path>,
+    faction_config: &FactionConfig,
+    resource_config: &ResourceConfig,
+) -> Result<(), Box<dyn Error>> {
+    let universe = load_map(config_path)?;
+    load_map_into_db(pool, &universe).await?;
+    load_factions_into_db(pool, faction_config).await?;
+    load_resources_into_db(pool, resource_config).await?;
+    if let Some(config_dir) = config_path {
+        let entity_configs = load_entity_configs(config_dir)?;
+        load_entities_into_db(pool, &universe, &entity_configs).await?;
+    }
+    Ok(())
+}
 
 pub async fn should_auto_load(pool: &SqlitePool) -> Result<bool, Box<dyn Error>> {
     let value = server_state_repo::get(pool, LAST_MAP_LOAD_KEY).await?;
