+++ src/game/map.rs
@@ -1,9 +1,7 @@
 pub mod universe;
-pub mod universe_config_sync;
 
 pub use universe::Universe;
 pub use universe::dungeon::Dungeon;
 pub use universe::navigation::Navigation;
 pub use universe::room::Room;
 pub use universe::world::World;
-pub use universe_config_sync::sync_universe_config;
